+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 ### Changed
 - systemtests: fix plugin postgresql testrunners, old postgres `bc` call [PR #1948]
 - stored: reserve/acquire a device on first incoming write data [PR #1951]
+- stored: add dedupable backend [PR #1955]
 
 ### Fixed
 - fix include-ordering on FreeBSD that could cause build issues [PR #1973]
@@ -547,5 +548,6 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 [PR #1940]: https://github.com/bareos/bareos/pull/1940
 [PR #1948]: https://github.com/bareos/bareos/pull/1948
 [PR #1951]: https://github.com/bareos/bareos/pull/1951
+[PR #1955]: https://github.com/bareos/bareos/pull/1955
 [PR #1973]: https://github.com/bareos/bareos/pull/1973
 [unreleased]: https://github.com/bareos/bareos/tree/master